mathclaw/pyproject.toml
[project]
name = "mathclaw"
version = "1.0.0"
description = "Autonomous Mathematical Discovery Engine"
readme = "README.md"
requires-python = ">=3.9"
license = {text = "MIT"}
authors = [
    {name = "MathClaw Project"}
]
keywords = ["mathematics", "ai", "discovery", "theorem", "proof", "symbolic"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Mathematics",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

dependencies = [
    "sympy>=1.12",
    "click>=8.0",
    "numpy>=1.20",
]

[project.optional-dependencies]
openai = ["openai>=1.0"]
anthropic = ["anthropic>=0.20"]
gemini = ["google-generativeai>=0.5"]
all = [
    "openai>=1.0",
    "anthropic>=0.20",
    "google-generativeai>=0.5",
    "pyyaml>=6.0",
]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "black>=23.0",
    "mypy>=1.0",
]

[project.scripts]
mathclaw = "mathclaw.cli:main"

[project.urls]
Homepage = "https://github.com/mathclaw/mathclaw"
Documentation = "https://mathclaw.readthedocs.io"
Repository = "https://github.com/mathclaw/mathclaw"

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["mathclaw*"]

[tool.black]
line-length = 100
target-version = ['py39', 'py310', 'py311', 'py312']

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
